feat: OPA TP, protocol and skeletal CLI
An OPA policy and key registration transaction processor, tested at the
protobuf level. Similar in design to the chronicle TP, but should have
no dependencies on the rest of the Chronicle project

Negative / property testing required for cases that cannot be created
with the client message builders.

References:

CHRON-201 CHRON-183

Signed-off-by: Ryan Roberts <ryan@blockchaintp.com>