fix(bootstrap): add embedded default-allow opa rule devmode option
Ahead of linking up the various Auth elements under development, for developer and user convenience working on Chronicle's main branch, in devmode initialize the opa-executor in the GraphQL bootstrap with an embedded default allow OPA policy and appropriate entrypoint if no --opa-rule and --opa-entrypoint args are provided. Users in devmode can override this feature by providing --opa-rule and --opa-entrypoint cli args for a given OPA policy.

Signed-off-by: Joseph Livesey <joseph.livesey@btp.works>